Fixed some of the scoreboard issues, improved character movement
Made the jump feel less floaty while still keeping the glide, and attempted to resolve the scoreboard cursor issue, but there are still weird glitches I can't figure out. I'll take another look later.

+++ Group/Game Project/Assets/Scripts/UI/Level/LevelCompleteUI.cs
@@ -67,8 +67,8 @@ namespace UI.Level {
 				ScoreboardEntryUI ui = entry.GetComponent<ScoreboardEntryUI>();
 //				int score = player.GetComponent<PlayerDataForClients>().GetScore();
 				int score = player.GetComponent<PlayerDataForClients>().GetTotalNutTime();
-				if (score == 0 || player.GetComponent<PlayerDataForClients>().GetScore < 2){
-					score = float.MaxValue;
+				if (score == 0 || player.GetComponent<PlayerDataForClients>().GetScore() < 2){
+					score = int.MaxValue;
 				}
 
 				ui.SetName(player.GetComponent<PlayerDataForClients>().GetName());
